Eradicate boost headers from hyrise headers
* Move type switches (which ultimately depend on boost::mpl) into
  implementation files
* Move all type_switch related logic into meta_storage.h
