+++ src/lib/access/AggregateFunctions.h
@@ -2,108 +2,13 @@
 #ifndef SRC_LIB_ACCESS_AGGEGATEFUNCTIONS_H_
 #define SRC_LIB_ACCESS_AGGEGATEFUNCTIONS_H_
 
+#include <vector>
+
 #include <storage/AbstractTable.h>
 #include <storage/HashTable.h>
 #include <storage/storage_types.h>
-#include <storage/meta_storage.h>
+
 #include <json.h>
-#include <vector>
-
-
-
-
-/* helper construct to avoid excessive use
- * of switch case in executePlanOperation
- * uses templated type_switch in src/lib/storage/meta_storage.h
- * and calls the the correct templated operator implemented
- * below
- */
-namespace hyrise {
-namespace storage {
-struct sum_aggregate_functor {
-  typedef void value_type;
-
-  const hyrise::storage::c_atable_ptr_t& input;
-  hyrise::storage::atable_ptr_t& target;
-  pos_list_t *rows;
-  field_t sourceField;
-  std::string targetColumn;
-  size_t targetRow;
-
-  sum_aggregate_functor(const hyrise::storage::c_atable_ptr_t& i,
-                        hyrise::storage::atable_ptr_t& t,
-                        pos_list_t *forRows,
-                        field_t sourceF,
-                        std::string column,
-                        size_t toRow): input(i), target(t), rows(forRows), sourceField(sourceF), targetColumn(column), targetRow(toRow) {};
-
-  template <typename R>
-  void operator()() {
-    R result = 0;
-
-    if (rows != nullptr) {
-      for (const auto& currentRow: *rows) {
-        result += input->getValue<R>(sourceField, currentRow);
-      }
-    } else {
-      size_t input_size = input->size();
-
-      for (size_t i = 0; i < input_size; ++i) {
-        result += input->getValue<R>(sourceField, i);
-      }
-    }
-
-    target->setValue<R>(targetColumn, targetRow, result);
-  }
-};
-
-struct average_aggregate_functor {
-  typedef void value_type;
-
-  const hyrise::storage::c_atable_ptr_t& input;
-  hyrise::storage::atable_ptr_t& target;
-  pos_list_t *rows;
-  field_t sourceField;
-  std::string targetColumn;
-  size_t targetRow;
-
-  average_aggregate_functor(const hyrise::storage::c_atable_ptr_t& i,
-                            hyrise::storage::atable_ptr_t& t,
-                            pos_list_t *forRows,
-                            field_t sourceF,
-                            std::string column,
-                            size_t toRow): input(i), target(t), rows(forRows), sourceField(sourceF), targetColumn(column), targetRow(toRow) {};
-
-  template <typename R>
-  void operator()();
-};
-
-template<typename R>
-void average_aggregate_functor::operator()() {
-  R sum = 0;
-  int count = 0;
-
-  if (rows != NULL) {
-    for (const auto& currentRow: *rows) {
-      sum += input->getValue<R>(sourceField, currentRow);
-    }
-    count = rows->size();
-  } else {
-    size_t input_size = input->size();
-
-    for (size_t i = 0; i < input_size; ++i) {
-      sum += input->getValue<R>(sourceField, i);
-    }
-    count = input->size();
-  }
-
-  target->setValue<float>(targetColumn, targetRow, ((float)sum / count));
-}
-
-template<>
-void average_aggregate_functor::operator()<std::string>();
-}
-}
 
 struct AggregateFunctions {
   enum type {
@@ -164,12 +69,7 @@ class SumAggregateFun: public AggregateFun {
    * if rows == nullptr the functor is executed
    * on all rows of the input table
    */
-  virtual void processValuesForRows(const hyrise::storage::c_atable_ptr_t& t, pos_list_t *rows, hyrise::storage::atable_ptr_t& target, size_t targetRow) {
-
-    hyrise::storage::sum_aggregate_functor fun(t, target, rows, _field, columnName(t->nameOfColumn(_field)), targetRow);
-    hyrise::storage::type_switch<hyrise_basic_types> ts;
-    ts(_datatype, fun);
-  }
+  virtual void processValuesForRows(const hyrise::storage::c_atable_ptr_t& t, pos_list_t *rows, hyrise::storage::atable_ptr_t& target, size_t targetRow);
 
   virtual DataType getType() const {
     return _datatype;
@@ -203,17 +103,7 @@ class CountAggregateFun: public AggregateFun {
    * if map_range_t rows == nullptr all values
    * are considered for counting.
    */
-  virtual void processValuesForRows(const hyrise::storage::c_atable_ptr_t& t, pos_list_t *rows, hyrise::storage::atable_ptr_t& target, size_t targetRow) {
-    size_t count;
-
-    if (rows != nullptr) {
-      count = rows->size();
-    } else {
-      count = t->size();
-    }
-
-    target->setValue<hyrise_int_t>(columnName(t->nameOfColumn(_field)), targetRow, count);
-  }
+  virtual void processValuesForRows(const hyrise::storage::c_atable_ptr_t& t, pos_list_t *rows, hyrise::storage::atable_ptr_t& target, size_t targetRow);
 
   virtual DataType getType() const {
     return IntegerType;
@@ -239,11 +129,7 @@ class AverageAggregateFun: public AggregateFun {
    * if rows == NULL the functor is executed
    * on all rows of the input table
    */
-  virtual void processValuesForRows(const hyrise::storage::c_atable_ptr_t& t, pos_list_t *rows, hyrise::storage::atable_ptr_t& target, size_t targetRow) {
-    hyrise::storage::average_aggregate_functor fun(t, target, rows, _field, columnName(t->nameOfColumn(_field)), targetRow);
-    hyrise::storage::type_switch<hyrise_basic_types> ts;
-    ts(_datatype, fun);
-  }
+  virtual void processValuesForRows(const hyrise::storage::c_atable_ptr_t& t, pos_list_t *rows, hyrise::storage::atable_ptr_t& target, size_t targetRow) ;
 
   virtual DataType getType() const {
     return FloatType;
