+++ src/lib/stdlib.hpp
@@ -15,9 +15,6 @@
 #include <iostream>
 #include <atomic>
 #include <thread>
-#include "boost/mpl/vector.hpp"
-#include "boost/mpl/map.hpp"
-#include "boost/lexical_cast.hpp"
-#include "gtest/gtest.h"
+//#include "gtest/gtest.h"
 
 #endif
